fold the measurement tests into the wheel test file
